Feature/bpm project dealer role

+++ BPM.Web.Dealer.UI/src/app/components/cart/cart.component.ts
@@ -3,19 +3,23 @@ import { CartService } from '../../services/cart.service';
 import { CartItem } from '../../models/cart-item';
 import { CommonModule } from '@angular/common';
 import { PurchaseOrderService } from '../../services/purchase-order.service';
+import { RouterLink } from '@angular/router';
+import { ToastrService } from '@iqx-limited/ngx-toastr';
 
 @Component({
   selector: 'app-cart',
-  imports: [CommonModule],
+  imports: [CommonModule, RouterLink],
   templateUrl: './cart.component.html',
   styleUrl: './cart.component.css',
 })
 export class CartComponent implements OnInit {
   cartItems: CartItem[] = [];
+  isPlacingOrder = false;
 
   constructor(
     private cartService: CartService,
     private purchaseOrderService: PurchaseOrderService,
+    private toasterService: ToastrService,
   ) {}
 
   ngOnInit(): void {
@@ -23,8 +27,17 @@ export class CartComponent implements OnInit {
   }
 
   removeItem(drugId: string): void {
+    const confirmed = confirm('Remove this medicine from the cart?');
+
+    if (!confirmed) {
+      return;
+    }
+
     this.cartService.removeFromCart(drugId);
     this.cartItems = this.cartService.getCartItems();
+    console.log('Toast Called');
+
+    this.toasterService.success('Medicine removed from the cart', 'Success');
   }
 
   increase(drugId: string): void {
@@ -54,6 +67,7 @@ export class CartComponent implements OnInit {
   }
 
   placeOrder(): void {
+    this.isPlacingOrder = true;
     const auth = JSON.parse(localStorage.getItem('AuthenticatedUserResponse')!);
     console.log(auth.authenticateResponseDto);
 
@@ -82,11 +96,17 @@ export class CartComponent implements OnInit {
     this.purchaseOrderService.createPurchaseOrder(request).subscribe({
       next: (response) => {
         console.log(response);
-        alert('Purchase Order Created Successfully');
+        this.isPlacingOrder = false;
+        console.log('Toast called ');
+
+        this.toasterService.success('Purchase Order Created Successfully');
+        this.cartService.clearCart();
+        this.cartItems = [];
       },
       error: (error) => {
         console.error(error);
-        alert('Failed to create Purchase Order');
+        this.isPlacingOrder = false;
+        this.toasterService.error('Failed to create Purchase Order');
       },
     });
   }
